Hacky window-sized canvas component
I don't really like the way the invalidation signal works here, it's weird to
pass in the draw function to the Canvas constructor (which, hm, should Canvas
even really be a _class_?). Perhaps an `Observable[Unit]` stream of
invalidations? Which could be spliced into all the view updates. In the style
of
https://medium.com/@mark_dj/a-little-scala-js-experiment-4a8da7b8ab8c#.d7ujiglww
perhaps.